% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_stanFootMulti.R
\name{print.stanFootMulti}
\alias{print.stanFootMulti}
\title{Print method for stanFootMulti objects}
\usage{
\method{print}{stanFootMulti}(x, ...)
}
\arguments{
\item{x}{A stanFootMulti object returned by stan_foot_multi().}

\item{...}{Unused.}
}
\description{
Thiss will provide a compact summary of the fitted footBayesX model.
}
